+++ Dockerfile
@@ -8,4 +8,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 
-CMD [ "python", "-m", "src.main", "run" ]
+CMD [ "python", "-m", "src.app", "run" ]
