+++ src/app/__main__.py
@@ -1,3 +1,4 @@
+from .app import main
+
 if __name__ == "__main__":
-    from .main import main
     main()
